Add it and describe
It is an alias for test. Describe allows to group tests. The current implementation has some known limitations. First, we have to choose between function describe behaving as expected or providing support for testing async functions. Second, the error doesn't show the line where the code fails, but the one where the error is thrown.